refactor: #17 validate.util로 함수 이동 및 중복 로직 하나로 refactoring
