+++ src/lib/utils/post.util.ts
@@ -49,64 +49,3 @@ export const generateFileName = (file: File): string => {
   const fileName = file.name.replace(/[^a-zA-Z0-9-_\.]/g, '');
   return `${timestamp}-${random}-${fileName}`;
 };
-
-// 일단 post 관련 함수는 여기에 모아 두고 나중에 다른 곳에 합치겠습니다.
-
-/**
- * 챌린지 게시물의 필수 입력값을 검증하는 유틸리티 함수
- * @param {ChallengePost} challenge - 검증할 챌린지 게시물 데이터
- * @returns {string | null} - 검증 결과 (오류 메시지 또는 null)
- */
-export const validateChallengePost = (challenge: ChallengePost): string | null => {
-  const { title, description, startDate, finishDate, category, executeDays } = challenge;
-
-  if (!title) return FETCH_MESSAGES.TITLE_REQUIRED;
-  if (!description) return FETCH_MESSAGES.DESCRIPTION_REQUIRED;
-  if (!startDate) return FETCH_MESSAGES.START_DATE_REQUIRED;
-  if (!finishDate) return FETCH_MESSAGES.FINISH_DATE_REQUIRED;
-  if (!category) return FETCH_MESSAGES.CATEGORY_REQUIRED;
-  if (executeDays.length === 0) return FETCH_MESSAGES.EXECUTE_DAYS_REQUIRED;
-
-  // 제목 길이 (30자 이내)
-  if (title.length > 30) {
-    return FETCH_MESSAGES.TITLE_TOO_LONG;
-  }
-
-  // 소개 길이 (500자 이내)
-  if (description.length > 500) {
-    return FETCH_MESSAGES.DESCRIPTION_TOO_LONG;
-  }
-
-  return null;
-};
-
-/**
- * 파일 형식 및 크기를 검증하는 유틸리티 함수
- * @param {File} file - 검증할 파일
- * @param {string[]} allowedTypes - 허용되는 파일 형식
- * @param {number} maxSize - 허용되는 최대 파일 크기
- * @returns {string | null} - 오류 메시지 또는 null
- */
-export const validateFile = (file: File, allowedTypes: string[], maxSize: number): string | null => {
-  // 파일 형식
-  if (!allowedTypes.includes(file.type)) {
-    return FETCH_MESSAGES.IMAGE_TYPE_INVALID;
-  }
-
-  // 파일 크기
-  if (file.size > maxSize) {
-    return FETCH_MESSAGES.IMAGE_SIZE_TOO_LARGE;
-  }
-
-  return null;
-};
-
-/**
- * 파일 유효성 검증하는 유틸리티 함수
- * @param {File} file - 검사할 파일
- * @returns {string | null} - 에러 메시지 또는 null(유효한 경우)
- */
-
-export const validateUploadFile = (file: File): string | null => {
-  return validateFile(file, FILES.ALLOWED_TYPES, FILES.MAX_SIZE);
-};
